+++ src/v/bytes/fwd.h
@@ -10,12 +10,5 @@
  */
 
 #pragma once
-#include "model/record.h"
-#include "storage/segment_appender.h"
 
-namespace storage {
-
-ss::future<>
-write(segment_appender& appender, const model::record_batch& batch);
-
-} // namespace storage
+class iobuf;
