+++ src/v/bytes/tests/iobuf_utils_tests.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/bytes_iobuf.h"
 #include "bytes/tests/utils.h"
 
 #include <seastar/testing/thread_test_case.hh>
