+++ src/v/compression/internal/snappy_java_compressor.h
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "bytes/iobuf.h"
+#include "bytes/fwd.h"
 
 namespace compression::internal {
 struct snappy_java_compressor {
