+++ src/v/compression/snappy_standard_compressor.h
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "bytes/iobuf.h"
+#include "bytes/fwd.h"
 
 namespace compression {
