+++ src/v/http/iobuf_body.h
@@ -11,7 +11,6 @@
 
 #pragma once
 
-#include "bytes/bytes.h"
 #include "bytes/iobuf.h"
 #include "seastarx.h"
