+++ src/v/kafka/protocol/offset_commit.h
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "bytes/iobuf.h"
+#include "bytes/fwd.h"
 #include "kafka/protocol/errors.h"
 #include "kafka/protocol/schemata/offset_commit_request.h"
 #include "kafka/protocol/schemata/offset_commit_response.h"
