+++ src/v/storage/segment_appender.cc
@@ -9,6 +9,7 @@
 
 #include "storage/segment_appender.h"
 
+#include "bytes/iobuf.h"
 #include "config/configuration.h"
 #include "likely.h"
 #include "storage/chunk_cache.h"
